Fix compiler warnings where keysyms are printed.
%lX expectes unsigned long which KeySym is an alias for.

+++ src/autotype/x11/AutoTypeX11.cpp
@@ -655,7 +655,7 @@ void AutoTypePlatformX11::SendKeyPressedEvent(KeySym keysym)
     int keycode;
 
     if (keysym == NoSymbol) {
-        qWarning("No such key: keysym=0x%lX", static_cast<long>(keysym));
+        qWarning("No such key: keysym=0x%lX", keysym);
         return;
     }
 
@@ -682,7 +682,7 @@ void AutoTypePlatformX11::SendKeyPressedEvent(KeySym keysym)
     /* determine keycode and mask for the given keysym */
     keycode = GetKeycode(keysym, &wanted_mask);
     if (keycode < 8 || keycode > 255) {
-        qWarning("Unable to get valid keycode for key: keysym=0x%lX", static_cast<long>(keysym));
+        qWarning("Unable to get valid keycode for key: keysym=0x%lX", keysym);
         return;
     }
